feat(core): finish first draft of config files
