+++ packages/core/src/config/types.ts
@@ -4,6 +4,12 @@ import { Configuration } from "webpack";
 export type Mode = "development" | "production";
 export type Target = "module" | "es5" | "node";
 
+export interface Headers {
+  [key: string]: string;
+}
+
+export type Robots = string;
+
 export interface BabelConfigs {
   module: TransformOptions;
   es5: TransformOptions;
@@ -14,4 +20,16 @@ export interface WebpackConfigs {
   module: Configuration;
   es5: Configuration;
   node: Configuration;
+}
+
+export interface HeaderConfig {
+  static: Headers;
+  dynamic: Headers;
+}
+
+export interface FrontityConfig {
+  headers: HeaderConfig;
+  robots: Robots;
+  babel: BabelConfigs;
+  webpack: WebpackConfigs;
 }
